rename fields -> snapshotFields, for clarity
