+++ lib/snapshot.js
@@ -7,7 +7,7 @@ var moment = require('moment');
 var Promise = require('bluebird');
 
 var _constants = require('./constants');
-var _fields = require('./fields');
+var _fields = require('./snapshotFields');
 var _utils = require('./utils');
 var getCrumb = require('./yahooCrumb').getCrumb;
