+++ lib/snapshotFields.js
